Alteracoes nos mapeamentos Conta,Gasto,Orcamento e Recebimento e services para corrigir problemas:  - Nao excluia conta,gasto e recebimento.  - Gerava erro ao tentar excluir orçamento

+++ src/financeiro/model/service/OrcamentoService.java
@@ -1,9 +1,14 @@
 package financeiro.model.service;
 
 import java.util.Date;
+import java.util.List;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
+import javax.persistence.Query;
+
+import org.jboss.logging.Logger;
+import org.primefaces.component.log.Log;
 
 import financeiro.model.bean.Conta;
 import financeiro.model.bean.FinanceiroException;
@@ -16,12 +21,14 @@ public class OrcamentoService extends ServiceGeneric<Orcamento, Integer> {
 
 	@EJB
 	private RecebimentoService recebimentoService;
-	
+
 	@EJB
 	private ContaService contaService;
-	
+
 	@EJB
 	private GastoService gastoService;
+
+	private Logger log = Logger.getLogger(this.getClass());
 
 	public void recebe(Recebimento recebimento, Orcamento orcamento) {
 		try {
@@ -33,7 +40,7 @@ public class OrcamentoService extends ServiceGeneric<Orcamento, Integer> {
 			throw new FinanceiroException(e);
 		}
 	}
-	
+
 	public void adicionaConta(Conta conta, Orcamento orcamento) {
 		try {
 			conta.setOrcamento(orcamento);
@@ -44,29 +51,71 @@ public class OrcamentoService extends ServiceGeneric<Orcamento, Integer> {
 			throw new FinanceiroException(e);
 		}
 	}
-	
+
+	/** 
+	 * necessario carregamento dos gastos do orçamento 
+	 * em virtudo do cascade.all e pelo fato do relacionamento ser bidirecional
+	 * sem o carregamento, só o gastoService.remove(gasto) nao apaga o gasto
+	 */
 	public void cancelaConta(Integer idConta, Orcamento orcamento) {
 		try {
 			Conta conta = contaService.encontra(idConta, Conta.class);
+			orcamento = this.carregaContas(orcamento.getId());
+			
+			if (orcamento.getContas()!=null && orcamento.getContas().size() > 0 ) {
+				orcamento.getContas().remove(conta);
+			} else {
+				throw new FinanceiroException("Orçamento " + orcamento.getId() + 
+						" nao possui contas relacionadas");
+			}
+			
 			orcamento.cancelaConta(conta);
 			contaService.remove(conta);
 			this.atualiza(orcamento);
+			log.info(">>> Cancelada conta " + idConta);
 		} catch (Exception e) {
 			throw new FinanceiroException(e);
 		}
 	}
-	
+
+	private Orcamento carregaContas(Integer id) {
+		Query query = this.entityManager.createNamedQuery("Orcamento.loadContas");
+		query.setParameter("id", id);
+		return (Orcamento) query.getSingleResult();
+	}
+
+	/** 
+	 * necessario carregamento dos gastos do orçamento 
+	 * em virtudo do cascade.all e pelo fato do relacionamento ser bidirecional
+	 * sem o carregamento, só o gastoService.remove(gasto) nao apaga o gasto
+	 */
 	public void cancelaGasto(Integer idGasto, Orcamento orcamento) {
 		try {
 			Gasto gasto = gastoService.encontra(idGasto, Gasto.class);
 			orcamento.cancelaGasto(gasto);
-			gastoService.remove(gasto);
+			orcamento = carregaGastos(orcamento.getId());
+			
+			if (orcamento.getGastos()!=null && orcamento.getGastos().size() > 0 ) {
+				gastoService.remove(gasto);
+			} else {
+				throw new FinanceiroException("Orçamento "+ orcamento.getId() + 
+						" nao possui gastos relacionados") ; 
+			}
+			
 			this.atualiza(orcamento);
+			log.info(">>> Cancelado gasto " + idGasto);
 		} catch (Exception e) {
 			throw new FinanceiroException(e);
 		}
 	}
-	
+
+
+	private Orcamento carregaGastos(Integer idOrcamento) {
+		Query query = entityManager.createNamedQuery("Orcamento.loadGastos");
+		query.setParameter("id", idOrcamento);
+		return (Orcamento) query.getSingleResult();
+	}
+
 	public void adicionaGasto(Gasto gasto, Orcamento orcamento) {
 		try {
 			orcamento.adicionaGasto(gasto);
@@ -77,7 +126,7 @@ public class OrcamentoService extends ServiceGeneric<Orcamento, Integer> {
 			throw new FinanceiroException(e);
 		}
 	}
-	
+
 	public void pagaConta(Conta conta , Orcamento orcamento ) {
 		try {
 			orcamento.pagaConta(conta, conta.getValorPago(), conta.getDataPagamento());
@@ -87,7 +136,7 @@ public class OrcamentoService extends ServiceGeneric<Orcamento, Integer> {
 			throw new FinanceiroException(e);
 		}
 	}
-	
+
 	public void efetuaPagamentoGasto(Gasto gasto, double valor,Date data) {
 		try {
 			Orcamento orcamento = gasto.getOrcamento();
@@ -98,16 +147,31 @@ public class OrcamentoService extends ServiceGeneric<Orcamento, Integer> {
 			throw new FinanceiroException(e);
 		}
 	}
-	
+
+	/** 
+	 * na exclusao do recebimento não é necessária a carga da lista de recebimentos 
+	 * pois o fetch do relacionamento é EAGER
+	 */
 	public void cancelaRecebimento(Integer idRecebimento,Orcamento orcamento) {
 		try {
 			Recebimento recebimento = recebimentoService.encontra(idRecebimento, Recebimento.class);
+			recebimento.setOrcamento(null);
+			
+			if (orcamento.getRecebimentos()!=null && orcamento.getRecebimentos().size() > 0 ) {
+				orcamento.getRecebimentos().remove(recebimento);
+			} else {
+				throw new FinanceiroException("Orcamento " + orcamento.getId() + 
+						" nao possui recebimentos associados");
+			}
+			
 			orcamento.cancelaRecebimento(recebimento.getValor());
 			recebimentoService.remove(recebimento);
 			this.atualiza(orcamento);
+			log.info(">>> Recebimento " + idRecebimento + " cancelado com sucesso");
 		} catch(Exception e ) {
 			throw new FinanceiroException(e);
 		}
 	}
-	
+
+
 }
